Refactor caching to separate utility
